+++ src/factory/index.ts
@@ -2,7 +2,7 @@ import * as isCi from 'is-ci';
 import stricter from '../stricter';
 import getDebugLogger from '../logger/get-debug-logger';
 import getNullLogger from '../logger/get-null-logger';
-import type { CliOptions, Stricter } from '../types';
+import type { CliOptions, Stricter, StricterArguments } from '../types';
 import getReporter from './get-reporter';
 import getConfigLocation from './get-config-location';
 import getCacheManager from './get-cache-manager';
@@ -12,16 +12,17 @@ export default (options: CliOptions): Stricter => {
     const reporter = getReporter(options.reporter);
     const logger = isCi ? getNullLogger() : getDebugLogger();
     const cacheManager = getCacheManager();
+    const stricterArguments: StricterArguments = {
+        logger,
+        reporter,
+        cacheManager,
+        options: {
+            configPath,
+            clearCache: options.clearCache,
+            fix: options.fix,
+            rulesToVerify: options.rulesToVerify,
+        },
+    };
 
-    return () =>
-        stricter({
-            logger,
-            reporter,
-            cacheManager,
-            options: {
-                configPath,
-                clearCache: options.clearCache,
-                rulesToVerify: options.rulesToVerify,
-            },
-        });
+    return () => stricter(stricterArguments);
 };