+++ src/reporter/junit.test.ts
@@ -20,6 +20,7 @@ describe('junitReporter', () => {
                 errors: [],
                 warnings: [],
                 time: 1,
+                fixes: [],
             },
         });
         const output = logMock.mock.calls[0][0];
@@ -32,6 +33,7 @@ describe('junitReporter', () => {
                 errors: [],
                 warnings: ['rule2-warning'],
                 time: 2,
+                fixes: [],
             },
         });
         const output = logMock.mock.calls[0][0];
@@ -47,16 +49,19 @@ describe('junitReporter', () => {
                 ],
                 warnings: [],
                 time: 1,
+                fixes: [],
             },
             rule2: {
                 errors: [],
                 warnings: ['rule2-warning'],
                 time: 2,
+                fixes: [],
             },
             rule3: {
                 errors: ['rule3-error'],
                 warnings: ['rule3-warning'],
                 time: 3,
+                fixes: [],
             },
         });
         const output = logMock.mock.calls[0][0];