+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Dependencies bump
 - Drop node 8 support
 - Expose types from `index` (fixes #114)
+- Add a possibility for rules to provide fixes for violations; they can be applied via `stricter --fix` (fixes #107)
 
 ## 0.4.1 - Dogfooding
 - Support for the scoped plugins (#93 by @nadiam84)